feat: refactor type tests for appConfig middleware and update package.json scripts
