+++ packages/validator/transpile.test-d.ts
@@ -1,4 +1,5 @@
 import { expectType } from "tsd";
+import { transpileLocale, transpileSchema } from "./transpile";
 
 const schema = transpileSchema({ type: "object" }, {});
 expectType<any>(schema);
